Add CodeChallengeType and its accompanying types
This is the initial stab at a "template" for creating challenges. It includes an ExampleCodeChallenge complete with example entries and tests.

In a future commit I will convert the TwoSum challenge and entries to this style, followed by an implementation for the other two currently unimplemented challenges we have on deck.